requirements.txt
anyio==4.8.0
appnope==0.1.4
APScheduler==3.10.4
asttokens==3.0.0
boto3==1.35.91
botocore==1.35.91
certifi==2024.12.14
charset-normalizer==3.4.1
comm==0.2.2
debugpy==1.8.11
decorator==5.1.1
executing==2.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyTelegramBotAPI==4.22.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.3
pytz==2024.2
pyzmq==26.2.0
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
telebot==0.0.5
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
wcwidth==0.2.13
google-api-core==2.24.1
google-api-python-client==2.132.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.68.0
pandas==2.2.2
numpy==1.26.4